Add ability to add a "label" to back-ed up filesystem
This is to mitigate the issue, for example when the filesystems file
has:

```
/var/portage
/usr/local/portage
```

In this case we'd want to back up both filesystems to /var/backup/<hostname>/0/portage.
This change adds an optional `:label` to `filesystems`, for example:

```
/var/portage:var-portage
/usr/local/portage
```

In this example, the first filesystem will be backed up under
`/var/backup/<hostname>/0/var-portage` while the second would be
`/var/backup/<hostname>/0/portage` (the usual default).