purgebackups: catch up
I hate to do commits like this, but apparently the purgebackups script I
had on my "production" instance had a lot of changes from the source
repo and so this commit is to catch up with production. I made a few
other changes (mostly type annotation changes) along the way.
Nevertheless, this is the latest/greatest purgebackups.